requirements.txt
pyttsx3==2.90
Pillow==10.0.0 